+++ checks/mkDummySrcTests/default.nix
@@ -1,22 +1,46 @@
-{ linkFarmFromDrvs
+{ lib
+, linkFarmFromDrvs
 , mkDummySrc
 , runCommand
 }:
 
 let
-  cmpDummySrc = name: path:
+  cmpDummySrcRaw = name: input: expected:
     let
       dummySrc = mkDummySrc {
-        src = path + "/input";
+        src = input;
       };
     in
     runCommand "compare-${name}" { } ''
-      diff -r ${path + /expected} ${dummySrc}
+      echo ${expected} ${dummySrc}
+      diff -r ${expected} ${dummySrc}
       touch $out
     '';
+
+  cmpDummySrc = name: path:
+    let
+      expected = path + "/expected";
+      input = path + "/input";
+
+      # Regression test for https://github.com/ipetkov/crane/issues/46
+      filteredInput = lib.cleanSourceWith {
+        src = input;
+        filter = path: type:
+          let baseName = builtins.baseNameOf path;
+          in
+          type == "directory" || lib.any (s: lib.hasPrefix s (builtins.baseNameOf path)) [
+            "Cargo"
+            "config"
+          ];
+      };
+    in
+    [
+      (cmpDummySrcRaw name input expected)
+      (cmpDummySrcRaw "${name}-filtered" filteredInput expected)
+    ];
 in
-linkFarmFromDrvs "cleanCargoToml" [
+linkFarmFromDrvs "cleanCargoToml" (lib.flatten [
   (cmpDummySrc "single" ./single)
   (cmpDummySrc "single-alt" ./single-alt)
   (cmpDummySrc "workspace" ./workspace)
-]
+])